// src/types/DadosFamiliar.tsx
export default interface DadosFamiliar {
    name: string
    cpf: string
    phone: string
    email: string
    password: string
    gender: string
    scholarity: string
    workingPlace: string
    occupation: string
    workPhone: string
    income: number
    familiarKinship: string
    status: string
    livesWithPatient: boolean
}